# Conflicts:
#	src/main/java/com/example/ProjectStock/Controller/StockSearchHistoryController.java
#	src/main/java/com/example/ProjectStock/Service/StockSearchHistoryService.java

+++ src/main/java/com/example/ProjectStock/Repository/StockEntityRepository.java
@@ -18,25 +18,25 @@ public interface StockEntityRepository extends JpaRepository<Stock, Long> {
     List<Stock> findByDateBetween(Date startDate, Date endDate);
 
     // Method for retrieving a list of open values for a given ticker and date range
-    List<BigDecimal> findOpenByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
+    List<Double> findOpenByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of high values for a given ticker and date range
-    List<BigDecimal> findHighByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
+    List<Double> findHighByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of low values for a given ticker and date range
-    List<BigDecimal> findLowByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
+    List<Double> findLowByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of close values for a given ticker and date range
-    List<BigDecimal> findCloseByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
+    List<Double> findCloseByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of adjusted close values for a given ticker and date range
-    List<BigDecimal> findAdjustedCloseByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
+    List<Double> findAdjustedCloseByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of volume values for a given ticker and date range
     List<Long> findVolumeByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of dividend amount values for a given ticker and date range
-    List<BigDecimal> findDividendAmountByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
+    List<Double> findDividendAmountByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
 
     // Method for retrieving a list of split coefficient values for a given ticker and date range
     List<Double> findSplitCoefficientByTickerAndDateBetween(String ticker, Date startDate, Date endDate);
